refactor(web): move computeAlignment core to standalone method
Relates-to: #14480

This is being done to prepare for a fix to #14480, which is needed to better handle complex input transforms and tokenization scenarios.  The easiest way to consistently index the tokenized transforms... is to perform context alignment on the results of transform application.
