+++ web/src/engine/predictive-text/worker-thread/src/main/test-index.ts
@@ -10,7 +10,7 @@ export * as correction from './correction/index.js';
 export * from './model-helpers.js';
 export * as models from './models/index.js';
 export { ModelCompositor } from './model-compositor.js';
-export { tokenizeTransform, tokenizeTransformDistribution } from './correction/transform-tokenization.js';
+export * from './correction/transform-tokenization.js';
 export * from './predict-helpers.js';
 export { default as TransformUtils } from './transformUtils.js'
 export { default as LMLayerWorker } from './index.js'
