+++ web/src/test/auto/headless/engine/predictive-text/worker-thread/context/alignment-helpers.tests.ts
@@ -9,7 +9,7 @@
  */
 
 import { assert } from 'chai';
-import { EditOperation, getEditPathLastMatch, isSubstitutionAlignable } from '@keymanapp/lm-worker/test-index';
+import { computeAlignment, EditOperation, getEditPathLastMatch, isSubstitutionAlignable } from '@keymanapp/lm-worker/test-index';
 
 describe('getEditPathLastMatch', () => {
   it('returns the last match when no substitutions exist', () => {
@@ -115,5 +115,455 @@ describe('isSubstitutionAlignable', () => {
   it(`returns true:  'applesauce' => 'plesauce' (subs not allowed)`, () => {
     // The double-p adds a fun complication once the first gets dropped.
     assert.isTrue(isSubstitutionAlignable('applesauce', 'plesauce'));
+  });
+});
+
+
+describe('computeAlignment', () => {
+  it("properly matches and aligns when contexts match", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [...baseContext];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: 0,
+      matchLength: 5,
+      tailEditLength: 0,
+      tailTokenShift: 0
+    });
+  });
+
+  it("properly matches and aligns with applied-suggestion contexts", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'o'
+    ];
+    const newContext = [...baseContext];
+    newContext[4] = 'over';
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: 0,
+      matchLength: 4,
+      tailEditLength: 1,
+      tailTokenShift: 0
+    });
+  });
+
+  it("properly matches and aligns with applied-suggestion at start of context", () => {
+    const baseContext = [
+      'te'
+    ];
+    const newContext = [
+      'testing',
+      ' ',
+      ''
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false, true);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: 0,
+      matchLength: 0,
+      tailEditLength: 1,
+      tailTokenShift: 2
+    });
+  });
+
+  it("detects unalignable contexts - no matching tokens", () => {
+    const baseContext = [
+      'swift', 'tan', 'wolf', 'leaped', 'across'
+    ];
+    const newContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {canAlign: false});
+  });
+
+  it("detects unalignable contexts - too many mismatching tokens", () => {
+    const baseContext = [
+      'swift', 'tan', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {canAlign: false});
+  });
+
+  it("fails alignment for leading-edge word substitutions", () => {
+    const baseContext = [
+      'swift', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {canAlign: false});
+  });
+
+  it("fails alignment for small leading-edge word substitutions", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'sick', 'brown', 'fox', 'jumped', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {canAlign: false});
+  });
+
+  it("properly matches and aligns when lead token is modified", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'uick', 'brown', 'fox', 'jumped', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: 0,
+      matchLength: 5,
+      tailEditLength: 0,
+      tailTokenShift: 0
+    });
+  });
+
+  it("properly matches and aligns when lead token is removed", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'brown', 'fox', 'jumped', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: -1,
+      matchLength: 4,
+      tailEditLength: 0,
+      tailTokenShift: 0
+    });
+  });
+
+  it("properly matches and aligns when lead token is added", () => {
+    const baseContext = [
+      'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: 1,
+      matchLength: 4,
+      tailEditLength: 0,
+      tailTokenShift: 0
+    });
+  });
+
+  it("properly matches and aligns when lead tokens are removed and modified", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'ox', 'jumped', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: -2,
+      matchLength: 3,
+      tailEditLength: 0,
+      tailTokenShift: 0
+    });
+  });
+
+  it("properly matches and aligns when lead tokens are added and modified", () => {
+    const baseContext = [
+      'rown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: 1,
+      matchLength: 4,
+      tailEditLength: 0,
+      tailTokenShift: 0
+    });
+  });
+
+  it("properly matches and aligns when lead token is removed and tail token is added", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'brown', 'fox', 'jumped', 'over', 'the'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: -1,
+      matchLength: 4,
+      tailEditLength: 0,
+      tailTokenShift: 1
+    });
+  });
+
+  it("properly matches and aligns when lead token and tail token are modified", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'ove'
+    ];
+    const newContext = [
+      'uick', 'brown', 'fox', 'jumped', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: 0,
+      matchLength: 4, // we treat 'quick' and 'uick' as the same
+      tailEditLength: 1,
+      tailTokenShift: 0
+    });
+  });
+
+  it("properly matches and aligns when lead token and tail token are modified + new token appended", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'ove'
+    ];
+    const newContext = [
+      'uick', 'brown', 'fox', 'jumped', 'over', 't'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: 0,
+      matchLength: 4, // we treat 'quick' and 'uick' as the same
+      tailEditLength: 1,
+      tailTokenShift: 1
+    });
+  });
+
+  it("properly handles context window sliding backward", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'e', 'quick', 'brown', 'fox', 'jumped', 'ove'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: 1,
+      matchLength: 4, // we treat 'quick' and 'uick' as the same
+      tailEditLength: 1,
+      tailTokenShift: 0
+    });
+  });
+
+  it("properly handles context window sliding far backward", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'the', 'quick', 'brown', 'fox', 'jumped'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: 1,
+      matchLength: 4, // we treat 'quick' and 'uick' as the same
+      tailEditLength: 0,
+      tailTokenShift: -1
+    });
+  });
+
+  it("properly handles context window sliding farther backward", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'the', 'quick', 'brown', 'fox', 'jumpe'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: 1,
+      matchLength: 3, // we treat 'quick' and 'uick' as the same
+      tailEditLength: 1,
+      tailTokenShift: -1
+    });
+  });
+
+  it("fails alignment for mid-head deletion", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'quick', 'fox', 'jumped', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {canAlign: false});
+  });
+
+  it("fails alignment for mid-head insertion", () => {
+    const baseContext = [
+      'quick', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {canAlign: false});
+  });
+
+  it("fails alignment for mid-tail deletion", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'quick', 'brown', 'fox', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {canAlign: false});
+  });
+
+  it("fails alignment for mid-tail insertion", () => {
+    const baseContext = [
+      'quick', 'brown', 'fox', 'jumped', 'over'
+    ];
+    const newContext = [
+      'quick', 'brown', 'fox', 'jumped', 'far', 'over'
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {canAlign: false});
+  });
+
+  it("handles late-context suggestion application after backspace", () => {
+    const baseContext = [
+      'quick', ' ', 'brown', ' ', 'fox', ' ', 'jumped', ' ', 'oven', ' ', ''
+    ];
+    const newContext = [
+      'quick', ' ', 'brown', ' ', 'fox', ' ', 'jumped', ' ', 'over', ' ', ''
+    ];
+
+    const computedAlignment = computeAlignment(baseContext, newContext, false);
+    assert.deepEqual(computedAlignment, {
+      canAlign: true,
+      leadTokenShift: 0,
+      matchLength: 8,
+      tailEditLength: 3,
+      tailTokenShift: 0
+    });
+  });
+
+  it("handles sliding context-window scenarios", () => {
+    // // Explicitly-defined window, though it's not needed directly by the method.
+    // const config = {
+    //   leftContextCodePoints: 64,
+    //   rightContextCodePoints: 64
+    // };
+
+    const baseContext1 = [
+      // "ap" prefix not in actual view, but preserved by prior tokenization rounds.
+      "applesauce", " ", "and", " ", "orange", " ", "juice", " ", "don't", " ", "seem", " ",
+      "like", " ", "they'd", " ", "make", " ", "for", " ", "the", " ", "be"
+    ];
+
+    const incomingContext1 = [
+      "plesauce", " ", "and", " ", "orange", " ", "juice", " ", "don't", " ", "seem", " ",
+      "like", " ", "they'd", " ", "make", " ", "for", " ", "the", " ", "bes"
+    ];
+
+    // 66 chars above, vs a sliding window of length 64.
+    assert.equal(baseContext1.reduce((accum, curr) => accum + curr.length, 0), 66);
+    // Actual window + one newly-typed character
+    assert.equal(incomingContext1.reduce((accum, curr) => accum + curr.length, 0), 65);
+
+    assert.deepEqual(computeAlignment(baseContext1, incomingContext1, true), {
+      canAlign: true,
+      leadTokenShift: 0,
+      matchLength: 22,
+      tailEditLength: 1,
+      tailTokenShift: 0
+    });
+
+    // Our tokenization scheme remembers the full original word before any of it slid out of
+    // the context window.
+    const baseContext2 = [
+      "applesauce", " ", "and", " ", "orange", " ", "juice", " ", "don't", " ", "seem", " ",
+      //                                                                 +2    +1     +4
+      "like", " ", "they'd", " ", "make", " ", "for", " ", "the", " ", "best", " ", "brea"
+    ];
+
+    const incomingContext2 = [
+      // "plesauce" => "e":  -7 chars.
+      "e", " ", "and", " ", "orange", " ", "juice", " ", "don't", " ", "seem", " ",
+      "like", " ", "they'd", " ", "make", " ", "for", " ", "the", " ", "best", " ", "break"
+    ];
+
+    // 73 chars above, vs a sliding window of length 64.
+    assert.equal(baseContext2.reduce((accum, curr) => accum + curr.length, 0), 73);
+    // Actual window + one newly-typed character
+    assert.equal(incomingContext2.reduce((accum, curr) => accum + curr.length, 0), 65);
+
+    assert.deepEqual(computeAlignment(baseContext2, incomingContext2, true), {
+      canAlign: true,
+      leadTokenShift: 0,
+      matchLength: 24,
+      tailEditLength: 1,
+      tailTokenShift: 0
+    });
+
+    const baseContext3 = [
+      "applesauce", " ", "and", " ", "orange", " ", "juice", " ", "don't", " ", "seem", " ", "like", " ",
+      "they'd", " ", "make", " ", "for", " ", "the", " ", "best", " ", "break"
+    ];
+
+    const incomingContext3 = [
+      " ", "and", " ", "orange", " ", "juice", " ", "don't", " ", "seem", " ", "like", " ",
+      "they'd", " ", "make", " ", "for", " ", "the", " ", "best", " ", "breakf"
+    ];
+
+    // 74 chars above, vs a sliding window of length 64.
+    assert.equal(baseContext3.reduce((accum, curr) => accum + curr.length, 0), 74);
+    // Actual window + one newly-typed character
+    assert.equal(incomingContext3.reduce((accum, curr) => accum + curr.length, 0), 65);
+
+    assert.deepEqual(computeAlignment(baseContext3, incomingContext3, true), {
+      canAlign: true,
+      leadTokenShift: -1,
+      matchLength: 23,
+      tailEditLength: 1,
+      tailTokenShift: 0
+    });
   });
 });
